make sure the Discord client is available for all files that may need to import it
