+++ src/index.js
@@ -1,3 +1,5 @@
+// require the Discord index.js first, as it sets up the client
 require("./discord/index.js");
+
 require("./replays/index.js");
 require("./twitch/index.js");
